code refactored + moved to scala-2.13 folder
