+++ tutorial/src/main/scala-2.13/anagrams/Anagrams2.scala
