+++ tutorial/src/main/scala-2.13/referentialTransparency/FutureIsNotRT.scala
