+++ tutorial/src/main/scala-2.13/referentialTransparency/IOIsRT.scala
