+++ tutorial/src/main/scala-2.13/referentialTransparency/NotRefTransparent.scala
