+++ tutorial/src/main/scala-2.13/tutorial/examples01To07/Example01ForComprehensions.scala
@@ -1,10 +1,11 @@
-package tutorial.ch01ForComprehensions
+package tutorial.examples01To07
+
 import scala.concurrent.Future
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Await
 import scala.concurrent.duration._
 
-object Example01 extends util.App {
+object Example01ForComprehensions extends util.App {
 
   def compute(list1: List[Int], list2: List[Int]): List[(Int, Int)] =
     for {
