+++ tutorial/src/main/scala-2.13/tutorial/examples01To07/Example03IdentityMonad.scala
@@ -1,0 +1,11 @@
+package tutorial.examples01To07;
+
+import tutorial.libMyCats._
+
+object Example03IdentityMonad extends util.App {
+
+  println("----- Identity:")
+
+  val idResult = compute(42, 42)
+  println(idResult)
+}
