+++ tutorial/src/main/scala-2.13/tutorial/examples01To07/README06FunctionsAreValues.md
