+++ tutorial/src/main/scala-2.13/tutorial/examples08/DbReaderApp.scala
@@ -1,4 +1,4 @@
-package tutorial.ch08ReaderMonad
+package tutorial.examples08
 
 import scala.util.chaining._
