+++ tutorial/src/main/scala-2.13/tutorial/examples09/IOAppCompute.scala
@@ -1,8 +1,7 @@
-package tutorial.ch09IOMonad
+package tutorial.examples09
 
 import scala.util.chaining._
 
-import tutorial.libCompute.WithMyCats._
 import tutorial.libMyCats._
 
 object IOAppCompute extends util.App {
