+++ tutorial/src/main/scala-2.13/tutorial/libMyCats/Monad.scala
